Updated sotfs blacklist in ParamBank.cs
Removed "generatorlocation", "generatorparam", and "generatorregistparam" from the sotfs blacklist. These three are needed when working with things that can't easily be worked with in the map editor, like phantoms or npcs that get generated on the fly without having explicit models in the maps. Example: I needed to edit these params to make the Pursuer spawn in instead of the Forlorn.